requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-allauth==65.9.0
django-cors-headers==4.7.0
django-filter==25.1
pandas==2.3.0
openpyxl==3.1.5
Pillow==11.3.0
python-decouple==3.8